Reviewed statistics
Solved all relevant comments. This entailed adding a readData function and an internal vector for storage to handle the data reading
separately from the calculations. This made the getScores function unnecessary and is therefore removed.
The test_stats file is also updated with smart pointers and SetUpTestSuite, as well as compiling all test data into consts.